修正

+++ setup.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+# 環境変数の設定
+RUBY_VERSION="3.1.2"
+RAILS_VERSION="6.1.4"
+NODE_VERSION="v18.18.2"
+NODE_DIR="/usr/local/share/node"
+PHP_VERSION="8.2"
+MARIADB_VERSION="10.5"
+
 # 結果を格納する変数を初期化
 installation_results=""
 
@@ -8,8 +16,8 @@ sudo yum update -y
 if [ $? -ne 0 ]; then installation_results+="パッケージ情報の更新が失敗しました。\n"; fi
 
 # 必要なツールとライブラリのインストール
-sudo yum install -y git curl bzip2 gcc gcc-c++ make openssl-devel readline-devel zlib-devel libffi-devel
-sudo yum -y install patch libyaml-devel zlib zlib-devel libffi-devel make autoconf automake libcurl-devel sqlite-devel mysql-devel
+sudo yum install -y git curl bzip2 gcc gcc-c++ make openssl-devel readline-devel zlib-devel libffi-devel \
+                    patch libyaml-devel zlib zlib-devel make autoconf automake libcurl-devel sqlite-devel mysql-devel
 if [ $? -ne 0 ]; then installation_results+="必要なツールとライブラリのインストールが失敗しました。\n"; fi
 
 # Rubyのインストール (rbenvを使用)
@@ -17,34 +25,28 @@ if [ ! -d "$HOME/.rbenv" ]; then
   git clone https://github.com/rbenv/rbenv.git ~/.rbenv
   echo 'export PATH="$HOME/.rbenv/bin:$PATH"' >> ~/.bashrc
   echo 'eval "$(rbenv init -)"' >> ~/.bashrc
-  # ~/.bashrc の変更を反映
   source ~/.bashrc
 
-  # ruby-buildのインストール
   git clone https://github.com/rbenv/ruby-build.git ~/.rbenv/plugins/ruby-build
 fi
 if [ $? -ne 0 ]; then installation_results+="rbenvのインストールが失敗しました。\n"; fi
 
-# Ruby 3.1.2 のインストール
-if ! rbenv versions | grep -q '3.1.2'; then
-  rbenv install 3.1.2
-  rbenv global 3.1.2
+# Ruby のインストール
+if ! rbenv versions | grep -q $RUBY_VERSION; then
+  rbenv install $RUBY_VERSION
+  rbenv global $RUBY_VERSION
   rbenv rehash
   rbenv exec gem install bundler
 fi
-if [ $? -ne 0 ]; then installation_results+="Ruby 3.1.2のインストールが失敗しました。\n"; fi
+if [ $? -ne 0 ]; then installation_results+="Ruby $RUBY_VERSION のインストールが失敗しました。\n"; fi
 
-# Node.js 16 のバージョンとインストールディレクトリを設定
-NODE_VER=v18.18.2
-NODE_DIR=/usr/local/share/node 
-
-# Node.js 18 のインストールに必要なパッケージをインストール
+# Node.js のインストールに必要なパッケージをインストール
 sudo yum install -y wget tar gzip
 
-# Node.js 18 のダウンロード＆インストール
-wget -nv https://d3rnber7ry90et.cloudfront.net/linux-x86_64/node-${NODE_VER}.tar.gz
-tar -xf node-${NODE_VER}.tar.gz
-sudo mv node-${NODE_VER} ${NODE_DIR}
+# Node.js のダウンロード＆インストール
+wget -nv https://d3rnber7ry90et.cloudfront.net/linux-x86_64/node-${NODE_VERSION}.tar.gz
+tar -xf node-${NODE_VERSION}.tar.gz
+sudo mv node-${NODE_VERSION} ${NODE_DIR}
 
 # バイナリへのリンクを貼る
 sudo ln -s ${NODE_DIR}/bin/corepack /usr/local/bin/corepack
@@ -52,42 +54,51 @@ sudo ln -s ${NODE_DIR}/bin/node /usr/local/bin/node
 sudo ln -s ${NODE_DIR}/bin/npm /usr/local/bin/npm
 sudo ln -s ${NODE_DIR}/bin/npx /usr/local/bin/npx
 
-# yarn のインストール
-npm install -g yarn
+# ダウンロードしたアーカイブファイルを削除
+rm -f node-${NODE_VERSION}.tar.gz
 
-# Rails 6.1.4 のインストール
-gem install rails -v 6.1.4
-if [ $? -ne 0 ]; then installation_results+="Rails 6.1.4のインストールが失敗しました。\n"; fi
+# yarn のインストールを確認し、必要ならインストール
+if ! command -v yarn > /dev/null; then
+  npm install -g yarn
+  sudo ln -s ${NODE_DIR}/bin/yarn /usr/local/bin/yarn
+  if [ $? -ne 0 ]; then installation_results+="yarnのインストールが失敗しました。\n"; fi
+else
+  echo "yarn はすでにインストールされています。"
+fi
+
+# Rails のインストール
+gem install rails -v $RAILS_VERSION
+if [ $? -ne 0 ];then installation_results+="Rails $RAILS_VERSION のインストールが失敗しました。\n"; fi
 
 # Nokogiri のインストール
 gem install nokogiri -v 1.16.6 -- --use-system-libraries
 if [ $? -ne 0 ]; then installation_results+="Nokogiriのインストールが失敗しました。\n"; fi
 
-# # Composer のインストール
-# if ! command -v composer > /dev/null; then
-#   php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-#   php -r "if (hash_file('sha384', 'composer-setup.php') === 'dac665fdc30fdd8ec78b38b9800061b4150413ff2e3b6f88543c636f7cd84f6db9189d43a81e5503cda447da73c7e5b6') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
-#   php composer-setup.php
-#   php -r "unlink('composer-setup.php');"
-#   sudo mv composer.phar /usr/local/bin/composer
-# fi
-# if [ $? -ne 0 ]; then installation_results+="Composerのインストールが失敗しました。\n"; fi
-
-# PHP 8.2 のインストール
-sudo amazon-linux-extras enable php8.2
+# PHP のインストール
+sudo amazon-linux-extras enable php${PHP_VERSION}
 sudo yum install -y php-cli php-pdo php-mbstring php-mysqlnd php-json php-gd php-openssl php-curl php-xml php-intl 
-if [ $? -ne 0 ]; then installation_results+="PHP 8.2のインストールが失敗しました。\n"; fi
+if [ $? -ne 0 ]; then installation_results+="PHP $PHP_VERSION のインストールが失敗しました。\n"; fi
 
-# composer のインストール
-php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-php -r "if (hash_file('sha384', 'composer-setup.php') === 'dac665fdc30fdd8ec78b38b9800061b4150413ff2e3b6f88543c636f7cd84f6db9189d43a81e5503cda447da73c7e5b6') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
-php composer-setup.php
-php -r "unlink('composer-setup.php');"
-sudo mv composer.phar /usr/local/bin/composer
+# composer のインストールを確認し、必要ならインストール
+if ! command -v composer > /dev/null; then
+  php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+  php -r "if (hash_file('sha384', 'dac665fdc30fdd8ec78b38b9800061b4150413ff2e3b6f88543c636f7cd84f6db9189d43a81e5503cda447da73c7e5b6') === 'dac665fdc30fdd8ec78b38b9800061b4150413ff2e3b6f88543c636f7cd84f6db9189d43a81e5503cda447da73c7e5b6') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
+  php composer-setup.php
+  php -r "unlink('composer-setup.php');"
+  sudo mv composer.phar /usr/local/bin/composer
+  if [ $? -ne 0 ]; then installation_results+="Composerのインストールが失敗しました。\n"; fi
+else
+  echo "Composer はすでにインストールされています。"
+fi
 
-# MariaDB のインストール
-sudo amazon-linux-extras enable mariadb10.5
-sudo yum -y install mariadb
+# MariaDB のインストールを確認し、必要ならインストール
+if ! command -v mysql > /dev/null; then
+  sudo amazon-linux-extras enable mariadb${MARIADB_VERSION}
+  sudo yum -y install mariadb
+  if [ $? -ne 0 ]; then installation_results+="MariaDB $MARIADB_VERSION のインストールが失敗しました。\n"; fi
+else
+  echo "MariaDB はすでにインストールされています。"
+fi
 
 # ターミナルのプロンプト表示設定
 echo '#!/bin/sh' > /home/ec2-user/prompt.sh
@@ -110,5 +121,6 @@ if [ -z "$installation_results" ]; then
 else
   echo -e "$installation_results"
 fi
+
